A class' refs to own class vars should not report as refs to the class
Class vars can be referenced from outside the owning class, and we want
to see these references as references to the class when browsing class
references. However we don't want to see a class' own use of its class
variables as refs. They are references, but the effect is just to obscure
the interesting references.

Similarly refs to classes defined in Namespaces should not report as refs
to the Namespace. Again they are references, but only indirectly.
Attempting to browse for references to Core before this change will give
some idea of why the previous behaviour just didn't work in practice.