Roundtrip pre-namespace packages in original source format for 7.1 compat
This provides backwards/forwards compatibility for D7 packages in D8 by
maintaining the existing source format, without namespaces. As long as no
namespaced objects are added to the D7 format package while in D8, it
should be possible to reload a D7 package modified and saved from D8.
In practice this means that the package can only reference variables that
are bindable from Smalltalk using unqualified names in D8. Obviously the
variable also has to exist in D7 in the Smalltalk dictionary.

- Erroneous references to namespaced classes in D7 packages are detected
when an attempt is made to save the package. All global identifiers must be
bindable using only an unqualifed name from Smalltalk. If this is not the
case then either the object (usually a class) must be aliased into Smalltalk,
or the reference removed, or the package upgraded to the D8 (paxVersion =
2.1) format.
- When a D7 package (paxVersion = 1) is loaded into D8, any
PoolConstantsDictionary definitions are automatically upgraded to the new
SharedPool namespace model. When a D7 format package is saved from D8,
any SharedPool classes it contains are saved back in the
PoolConstantsDictionary format.
- Any references to namespaced classes through aliases should be
maintained in file outs from D8, including for the method class of
extension methods, superclasses, etc.
- If a view resource from a D7 format package is edited in D8, the
saved resource method will use D7 style class references (i.e. all classes
are specified as in Smalltalk). It is an error if a class appears in the
resource that is not bindable by unqualified name from Smalltalk, and
this should be detected when attempting to save the resource.

Assuming that care is taken when modifying code from D7 in D8 to not
introduce dependencies on D8 specifics, then the code should round trip.
Some particular cases to be aware of are:
- Most qualified references in code will not work in D7. This case is
not currently validated in D8. Any qualified name that refers to one of
the D8 namespaces will definitely not work.
- Binding reference literals in code will not work at all in D7. This case
is also not currently validated in D8.
- Some packages have been split up in D8, especially those in the MVP
area. This will result in incorrect pre-requisites for D7. I intend
to perform the same package refactorings in D7 to address this.
- Obviously any dependencies on classes or methods that do not exist
in D7 will not work. This isn't specific to namespaces, but is rather
a general forwards compatibility issue.