Revert "File out Burning River packages in new format"
This reverts commit 7adbd7f55e36849091c94cdea009395ffed3dc1c, although
there remain some changes in the ExternalProcess package. This adds
extension methods to KernelLibrary that reference structure classes from
the package that are defined in Smalltalk. Classes in Smalltalk are not
visible through unqualified names in Kernel, so these references have
to be fully qualified pending work to support  compiling individual
methods in a different namespace than that of the method class, which
in turn requires support for namespace annotations on methods.