Fix broken "abbreviations" keyboard commands
Abbreviations class var is present, but nil. Dolphin packages don't provide
a mechanism to extend a class with new class variables, so this has been
hacked in by using an imperative expression, the result of which is that
the var becomes part of the definition of SmalltalkWorkspace and is
included in its base definition. The hack isn't really necessary - it would
be better to add another class to own the abbreviations, and just hook to
that from extensions added to workspace. In the meantime, just patch it up.