Fix #1111: Add alias for URLMonLibrary and ensure rko281/GitHub loads
The GitHub package will save out largely unchanged in a format that should
round trip back to 7.1, except for the the AXTypeLibraryAnalyzer
global. The TLA instance will not deserialize correctly in 7.1 if saved
from 7.2 in legacy package format (with binary-filed non-source globals)
as the class has changed instance variable shape in 7.2.
Generally it is not necessary to retain the TLA instance, so it could
probably be removed along with the references from #initializeTypeLib
class methods on the interface classes.

No guarantee that the package works in 7.2 once loaded, however.