Add Compiler support for generic method annotations, and namespace:
Methods can now have any unary or keyword annotation. The arguments to
keyword annotations can be any literal, or static variables. The
annotations are modelled as a <sequencedReadableCollection> of <Message>.

The only two uses in the base system currently are the special cases
<mutable> (to mark inst var setter methods that can modify immutable
objects), and <namespace: NS> for compiling a method with a custom
environment explicitly. This is in addition to the mechanism for compiling
a method with an implicit custom namespace that is used (for example) to
compile the extension/loose methods of legacy packages in the Smalltalk
namespace for compatibility. Explicit method namespaces are most likely to
be useful in extension methods too, although there is also the option of
using qualified names for static variables.