Prevent ViewComposer idle update death spiral if composing view closed
The ViewComposer attempts to refresh at idle time, which can fail
attempting to redraw adornment if the composing view is closed for some
reason. This results in a high rate of walkback dialogs that cannot be
recovered from.

One reason that the composing view might be closed is that in Windows
10 it appears in the list of application windows. This needs to be fixed,
but the increase in robustness of adding a guard to prevent failure in
the idle time processing is worthwhile anyway.