Upgrade to Scintilla 5.0.3
This adds the element colour abstraction, which simplifies and generalises
the management of a number of colour settings. In some cases this also
allows for the use of translucent colours, such as for the selection
highlights.

Support has been added for the auto-completion list element colour
settings which will be required for any future work on supporting themes.

There are a number of other minor improvements to add/improve the Scintilla
projection.