Improve workspace variable tips
With the introduction of namespaces, there is an increase in the ambiguity
of static variable references in code, so it is useful to see the full
name to which an identifier will bind when browsing code, or in the
debugger. This is a good opportunity to improve the quality of the variable
tips in general, to provide additional information that might be useful
such as whether an identifier refers to a temporary variable, argument,
or instance variable.