Open dialogs on "active" monitor if there is no active Dolphin owner window
Adds the notion of an active DisplayMonitor, which is determined using a
heuristic designed to identify the current target of the user's attention.
Normally the active monitor will be that on on which the foreground Window
is displayed. The exception is when the foreground window is the Windows
'program manager' window, which indicates that that the desktop has focus,
in which case the monitor under the current mouse cursor position is
regarded as the active monitor.

Dialogs are opened on the active monitor when they are not provided with an
owner and Dolphin does not have an active window.

Tricky to test because applications cannot rely on being able to change the
foreground Window in more recent versions of Windows. In order to simulate
this, as FakeUserLibrary is introduced for testing purposes that overrides
the API calls used to query from Windows the active and foreground window
handles.

Note the breaking change that `View class>>active` now returns nil if there
is no active Dolphin window (i.e. another application has the foreground
window), rather than an invalid View instance with null handle. The same
is true for `View class>>foreground`