Fix regression caused by original fix for #1115
The regression is that in-place editor dialogs can open on the wrong
display when the owner window is split between monitors. It can be
reproduced by placing a View Composer so that it is split between screens,
but mainly on the screen to the left, and with the property inspector on
the monitor to the right. When one of the in-place editor dialogs is opened
from the property inspector, the dialog comes up on the left monitor. This
is because it gets repositioned on to the monitor with which the window is
associated, which is that monitor on which the window is mostly displayed.