Add MVP list support for adding a sequence of items as one operation
Inserting a number of items into an MVP list is rather slow, which is
especially noticeable when using the ListTreeView, since it inserts items
when a "node" is expanded. This is largely because the addition of multiple
items is achieved by individually inserting single items, causing a great
deal of redundant view updating. By collapsing the add into a single update
event we can greatly improve the performance of the operation. This does
mean extending the public events API of ListModel, which is a breaking
change for any custom List view implementations.