Improve DelayTest failure messages and increase overshot tolerance
This is to (probably) fix #1073:
- Rearrange code to avoid allocating objects in the timed sections so
reducing the chance of triggering a GC.
- Calculate the startTime as close as possible to waiting on the Delay,
and inside the forked process where applicable, in case there is a delay
in scheduling it (e.g. it's allocation causes a GC event, or some other
interrupt is delivered).
- Increase the maximum overshot slightly as 3mS is too tight given that
the maximum timer clock resolution is 1mS.
- Fix the failure messages so they actually report the expected delay
rather than 00:00:00mS.