Ensure that fileout of legacy packages maintains prereqs with Old Names
In order to support seamless loading of legacy packages from 7.1 and
earlier, before the addition of namespaces, a number of packages in the
system have 'Old Names' siblings that contain just a set of shared
variables in the Smalltalk namespace mapping to classes in other namespaces
that would not normally be bindable from Smalltalk.

As well as being able to load legacy packages into Dolphin 8, it is also
useful to be able to save them out without change, as this can allow (with
some restrictions) code to be shared between versions and worked on in
either Dolphin 8 or Dolphin 7.1. The majority of contributed packages in
Dolphin 8 have been deliberately maintained in the Dolphin 7.1 format for
this reason, making it much easier to port changes between the versions.

Where 'Old Names' packages were among the package pre-requisites, they
were removed from the pre-requisite list by the legacy format filer.
It is necessary to remove them, as they will not exist in Dolphin 7.1.
However, there was a gap in that the result would have the original
pre-requisite package removed. In order to maintain the legacy package
pre-requisites correctly, instead of just removing the 'Old Names'
pre-requisites, we need to emit the corresponding package for which it
contains aliases.