Add `Object>>value`
There are pros and cons to this, but overall it is just to useful to
continue to exclude. The particular use case that tipped the balance for
me is wanting to be able to use either a block (to defer calculation) or
a value as an argument. The main downside is that a class of logic errors
may now be detected later since things not intended to be used as
valuables will respond to `#value`.