Add Graphics.ColorAlpha, a colour that decorates another with alpha
Useful, for example, to add alpha to a system colour without losing the
themed nature of the colour (i.e. the colour will still track windows
theme changes). This was motivated by the desire to associate a
system colour with one of the colouralpha settings in Scintilla, but it
has broader applications.

Also improve ColorView so that it shows the colour blended with Color
window to take account of any alpha.