Add tests for `SmalltalkLocale>>displayDuration:on:` and make clearer
SmalltalkLocale>>displayDuration:on: contains a lot of magic numbers.
The intention is a bit more obvious if we use named durations and include
an explanation of what the various elements of the format descriptions
mean.
It can also be written a little more optimally without obscuring the code.
The test class was in the wrong namespace.