UnitTestPlugin runs whole class suite when inherited tests selected
The UTP is supposed to run individual tests that might be selected as
methods in the class browser. This doesn't work for inherited tests and
it ends up running the whole class suite if only inherited tests are
selected, or just the non-inherited tests if there is a mixed selection.