Add setters for common control custom draw modes
Various of the custom draw item modes can be changed in custom draw
handlers, although there is no real consistency across the common controls.
An example is the isDisabled (CDIS_DISABLED) state. SysListView32 respects
this and draws the item with grayed text, and does not draw a selection
hot highlight. SysTreeView32 does nothing.