Bug in SelectionChangeEvent where no selection represented as #(nil)
When a SelectionChangeEvent is sent and the previous selection state was
no selection, or the new selection state is no selection, the created
event should have empty old or new selections, respectively, but was
actually a #(nil).

Also add the multiple-selection get/set methods to the selectableItems
protocol and implement these in MoenTreeView for consistency.