Depending on focus, System Browser toggles may not work correctly
The PackagedClassSelector holds a ClassListSelector, both of which had
(separate) flags for tracking toggle state. Depending on where the focus
is when a toggle (toolbar or instance mode radio buttons) is selected, the
outcome could be different. We need just one set of flags, and to take
care that the toggle affects both hierarchy and by-namespace views for the
unqualified/qualified names toggle.