Implement initializer mechanism for Images
Images have always been able to load from files or module resources, but
the loading mechanism was not very clean or flexible. We can tidy this up
and allow for creation and (importantly) recreation of images from other
sources by taking inspiration from the Gdiplus package and implementing a
hierarchy of initializer classes for Images. I've wanted to do this for a
while, but finally went ahead to enable the 16x16 footer icon for the new
MessageBox to be rendered with the best available image, including from
TextTileIcons.

This also makes the IconWithExtent extension in the "SSW Widget Enhancements"
package obsolete, as the base icon class can now have a specified extent
for the image that it will load by default from set of sizes available.