Argument to #at:ifAbsentPut: no longer needs to be a block
Rewrite some cases where a block is no longer required.
Not safe in general where the last arg is a variable or the result of an
expression as we don't know that the object will not respond to value
to answer something other than itself. Where the type is known not to
override Object>>value, it is safe.

For literals the rewrite generally preserved behaviour.