Fix intermittent dangling-pointer bug in block activation
In rare cases an activated block may be freed when activated causing a
subsequent crash. The issue was that the block could be unreferenced at
the point where a context needed to be allocated for blocks with env temps.

The bug was reliably reproduced by running AbstractParserErrorTest
testErrorsBlockNestingTooDeep on a debug VM.