Fix extract method refactoring where source method has custom namespace
Extracting from a method with a custom namespace may need to preserve the
custom namespace of the method, either from the source method's package or
from its namespace annotation. If this is not done, then any static
variable references in the new method may not be bindable. The alternative
would be to rewrite the static variable references so that the names are
bindable from the class' own namespace. The latter solution is probably
most appropriate for Extract to Component, which has a similar problem
when the target class has a different namespace even when extracting from
a method that does not have a custom namespace (to be addressed separately)