Fix issues with method refactorings losing custom namespace/package
When performing method refactorings the source method may be a loose method
or in a custom namespace. If the relationship with the package is not
maintained, the resulting method(s) may not be correctly packaged. The
RB code has previously been modified to maintain the package, but it was
done in such a way that the package was added to the AddMethodChange after
its creation, which may now be too late. Now that packages can also have a
custom namespace, the code may fail to compile if not compiled directly
in the context of the package. Also methods can have custom namespaces
independently of packaging, which must also be maintained.

There are other cases, e.g. extract to component, where the static
variable references in a refactored method created from another may need
to be rewritten to account for visibility in the target context. Typically
this would involved expanding unqualified or partially qualified names to
fully qualified names. This work is still pending.