Add support for converting FILETIMEs to/from DateAndTime
FILETIME is the native Windows filesystem time format. Not all DateAndTimes
will round trip to a FILETIME and back because the FILETIME has range and
precision limitations. It measures the number of ticks since
1st Jan 1601 UTC using an unsigned 64-bit integer. Therefore it cannot
represent dates before that point, nor dates more than about 58073 years
into the future. There may also be some very small loss of precision
because a tick is 100 nanoseconds, and DateAndTime uses unlimited precision
fractional arithmetic internally so can represent more precise times. This
is unlikely to be a problem in practice, although it means care needs to
be taken with equality comparisons.