Add a proper Registry Serializer and employ for XMLPad sample settings
The serializer is general purpose (i.e. you can theoretically serialize
any object graph into the registry as registry keys and values), but there
are some important caveats:
1) Registry paths are a maximum of 255 characters, including the root key
name, e.g. HKEY_CURRENT_USER, which will probably be insufficient for any
complex object graph, depending on the depth of the graph, and the length
of class names and instance variable names.
2) Registry trees can only be up to 512 levels deep, although the prior
limitation is likely to bite beforehand.
3) Microsoft recommend that individual values should not be larger than
2048 bytes.
4) The serialization and deserialization is very slow compared to other
alternatives.
5) Serialization of code objects to/from the current user key is a
potential security risk. For this reason the types of objects that can be
deserialized is deliberately limited to a preconfigured set, although this
behaviour is configurable and overridable.
With the above in mind, the registry serialization is only really intended
for saving dictionaries of values, where some of those settings might
be other simple collections, e.g. arrays of objects with mainly value
types. It works well for application settings, for example. This is
demonstrated in the revisions to the XmlPad sample.