Fix/improve refactoring experience (part 1)
Refactoring has been disrupted significantly by the addition of namespaces.
Some work has been done to fix things up, but there are still a lot of
issues. This commit started off as an attempt to fix some problems when
extracting code from a method in a custom namespace - the extracted code
might not compile due to the binding scope changing if the namespace is
not preserved, or is different in the target class. The former case is
fixed, the latter is still pending.

Along the way a bunch of other changes were made to try and fix some of
the things that have annoyed me about the refactoring experience for a
while:
1) Repeated prompts, sometimes in a loop until the input is valid
2) Lack of clarity as to the impact of a choice when prompted
3) Lack of explanation as to why the Ok button is disabled in a
Prompter requesting input (a problem with validating Prompters more
generally)
4) Lack of information in warnings as to the relevant code
5) Unable to proceed with Remove Method refactoring when there are
senders i.e. override the warning and remove anyway, creating the need
for a separate 'Delete Method' command that is not a refactoring.

Summary of changes to try and address these things, although not completely
in all cases yet:

- Fix cases of method extracts failing to compile due to not maintaining
a custom namespace in the source method (from original complication adding
an implicit annotation, or from the default namespace of the owning
package). Further work is required to fix 'Extract to Component' where the
static variable references may need to be rewritten to be bindable in the
target component.
- Improve message prompts shown for refactoring options to use more of the
richness of the new MessageBox, such as custom buttons appropriate for the
question being asked instead of Yes/No. This clarifies the choice to be
made (2).
- Make better use of task dialog based MessageBox for refactoring errors (2)
- Improve messages shown for certain refactoring warnings such as when
renaming a method with a selector defined in other classes. The use of
RefactoringWarning is replaced with refactoring options for these cases.
This clarifies the issue being warned about (4)
- The improved warning message for the Remove Method refactoring that is
shown when the removal candidate is still sent now allows the user to
choose to proceed regardless. This makes the separate Method Delete
operation (that was not a refactoring) redundant, so it is removed from
the UI as a simplification.
- Provide validation hints in prompts such as for the initializer
expression when adding a parameter to a method. This provides some
explanation as to why the Ok button is not enabled in the prompter, which
could otherwise be difficult to understand (3). The hints are displayed using
MessageBubbles, and the mechanism is generic and available for any use of
the standard Prompter.
- In-place rename of variables in a workspace will also show validation
hint bubbles when an attempt is made to commit the rename and the
preconditions are not met. The validation is improved to detect the cases
of attempting to use a local variable name that is already in use in the
code. For some reason these do not form part of the preconditions for the
Rename Temporary Refactoring, but rather are detected as errors when
executing the refactoring. As these are now pre-validated, the errors
should not be encountered from in-place renames.
- Improve validation for certain refactorings that require user input such
as the initializer expression of Add Parameter by using the actual
Refactoring's own preconditions to validate in the prompter. This should
prevent situations where it was possible to provide input and press Ok,
but then still get an error from the Refactoring (1).
- Open a Method Explorer to browse references to a method to be removed
that is referenced, rather than an EnvironmentBrowser. The latter is
effectively a filtered class browser, and does not provide a experience
for browsing a list of individual methods.
of methods.