Change order of Remove Class refactoring preconditions to check refs last
Now that we allow the user to override the precondition that removed
classes should not be referenced, this condition should be last to avoid
prompting for references and then failing on another precondition anyway.

Also make sure the references used for the precondition are the same as
those shown in the prompt, and the browser if requested.

Lastly, class variable names have always been Strings in Dolphin, not
Symbols. We don't need to intern the class names in the RB model either
since Strings and Symbols compare equal.