Make the 'remove duplicate methods' warning from push-up an option
The push up method refactoring has a warnign that pushing up a method will
result in duplicate definitions in direct or indirect subclasses of the
superclass. By making this a refactoring option we can show a richer
message that includes the list of duplicates, and allow the user to choose
to remove, not remove, or cancel/abort.

Also change the displayString of all relevant refactorings to use
unqualified class names. Although this may result in some ambiguity, this
is not normally an issue and makes the labels shown in any prompts, or in
the undo history, easier to read.