Secure CRT functions that return errno_t can use FFI errno return code
The errno return code will fail the FFI call primitive if the external API
call returns a non-zero errno. This allows a CRTError exception to be
raised by the library, rather than all the callers having to test for
non-zero code. Less code, simpler, and more efficient.

Note the special case for the STRUNCATE errno. This is more of a status
code/warning than an error, indicating a buffer was of insufficient size
causing output to be truncated, and so is still returned rather than
throwing an exception.

Also includes:
- a minor optimisation of `String>>#subStrings`
- removal of old deprecated methods in CRTLibrary and KernelLibrary