Ellipsise text provided to common control callback if too large for buffer
The callbacks from common control to request the text to display for an
item provide a buffer in which the text is to be stored. This seems to be
limited to 260 characters (i.e. _MAX_PATH, which would have been suitable
for the Windows Explorer before the advent of long filenames). Rather than
just truncated off the string if it is too large, we should ellipsise it
so that it is clear to the user that truncation has taken place. The
Unicode ellipsis character can be used to minimise the text loss (one
UTF-16 code point rather than three)
rather than three)