Add validater to prompt for new class name from MoveMethodRefactoring
When moving a method to a component one has to first choose the target
variable, and that is used to deduce a set of likely target classes. If
these are not correct, one can manually add a target class. The manual add
experience can be improved by validating that the entered class name is
actually a class name by extending ChoicePrompter to support the new
validater/ValidationResult bubble feature, as opposed to showing a message
box from the . A further improvement would be to show a class selection prompt instead, but this requires an
enhancement to ChoicePrompter to allow the use of a custom prompter
instead of the standard Prompter. Better still would be to design a
specific dialog for this option.