Fix enter key not working in in-place edits in browsers after image restart
In some browsers, after restarting a saved image an attempt to perform
a variable rename in the source code would bring up the in-place text
editor as expected, but it would not respond to the Enter key. Causing
focus to be lost by another means, e.g. clicking on another window, would
cause the edit to be committed and a rename refactoring of the appropriate
sort to be performed. The cause of the inactive Enter key was difficult to
track down, but turned out to be because a default button was set in the
view. This causes the Enter key press to be consumed as a "dialog"
navigation action on the pre-translation side of the message loop, so it
is never dispatched. It could be reproduced by switching to the Unit Test
plugin and clicking the Run Tests button (making it the default), then
saving the image and restarting. When the image comes back up and the view
is restored, the PushButton restores with the default button style set, and
so it makes itself the default in the shell. When the tabs are then changed
(automatically without the previous focus bug fix), the button (now hidden)
is still the default and will effectively be consuming the return key
presses, triggering a WM_COMMAND for the button's command id. On
the whole this goes unnoticed because the command is not enabled, but it
could cause an unexpected action to occur. In order to avoid that, this fix
is made in the MVP framework to ensure that the default button does not
remain set on a hidden tab. The fix will also restore the previous default
button on a tab when it is reactivated, which did not happen before.