Refactor StSemanticAnalyzer so can define lightweight variable binder
In order to implement class rename refactorings correctly with namespaces
it will be very helpful to be able to bind parse trees to the static
variables to which they refer. While this can be done using the
StSemanticAnalyser, is does a lot more as it is intended for full
compilation. This commit introduces an abstract superclass for semantic
analysers that contains the basics.