Further work on fixing RenameClassRefactoring to work with Namespaces
Class renaming now works more correctly and efficiently:
1) Bind all variable references against the state of the model before
performing rename operations in the model so that the original binding
resolution for each static variable and BindingReference can be compared
against the resolution post rename in order to determine if the
qualification of names can or must be changed. This is achieved using a
specialised semantic analyser, RBVariableAnalyser. RBMethod's now cache
their parse tree.
2) Remove renaming of Symbols that are the same as the unqualified name of
the class being renamed. This was never guaranteed to be correct, and
certainly isn't now as the unqualified names can be ambiguous. Instead
convert some of the common indirect class reference patterns to equivalent
BindingReference expressions. A to-do is to .
3) Add validation hints to the Create Subclass dialog

Still to do:
- Rename variable references in the renamed class itself if its namespace
is changed since this may affect the binding resolution
- Issue a refactoring warning should there be references to the original
unqualified class name symbol
- Renaming namespaces requires rewriting references to members of the
namespace as well as direct references to the namespace