Add refactoring option to warn when symbolic class names found on rename
Now that unqualified class names are ambiguous we can no longer just
rename any occurence of the name symbol in code (although that was always
a bit unreliable anyway, since not all occurrences of the name were
guaranteed to be indirect references to the class being renamed, even
without namespaces). We attempt to replace some common indirect reference
patterns when renaming classes out of Smalltalk where we can be sure that
these were references to the class, e.g. `Smalltalk at: #Foo` was an
indirect reference to the class Foo (if it existed). In this case the
substitution would be `#{Smalltalk.Foo} value`. It isn't feasible to
identify all such cases, so there may be some remaining occurrences of #Foo
in the code that we are not sure about. We can use a refactoring option to
allow the user to decide how to proceed - go ahead and rename, browse the
classes and cancel for now, or abort.