Workspace in-place renames should be deferred to avoid wait cursor
When an in-place edit is initiated in a workspace, e.g. to rename a
variable, if the command is not deferred then the cursor will flash
between the text bar and wait cursor if the mouse is moved. This is
because the Cursor>>showWhile: that wraps commands by default is still
on the stack.

Also ensure that the Scintilla blinking caret is hidden during the rename
operations as otherwise it can sometimes break through.