Revert "Inline BlockClosure>>ifCurtailed: into BlockClosure>>ensure:"
BlockClosure>>ifCurtailed: cannot be inlined into BlockClosure>>ensure:
without modifying the VM far return primitive as it expects to find
the ifCurtailed frame, and returns to the wrong place when an error
occurs in an ensure block causing curtailment of another.

This reverts commit 13888030474f967f174c5759a80d3996ef3cd061, and adds a
regression test that detects the fault the reverted commit introduced.