Literal Arrays and ByteArrays should have separate node and token classes
The use of an 'isForByteArray' flag is a smell, and introducing separate
node and token classes does indeed simplify the code allowing removal of
a number of conditional tests. It should also provide some performance
benefit as we no longer need to make needless visits to the parse nodes
representing literal ByteArrays when performing semantic analysis.