Store literal class references in STL as binding references
Version 5 of STL makes the literal arrays true static literals that don't
require the use of ##() expressions to represent as source text. This is
achieved by using literal binding references to refer to classes.

To help with forwards compatibility, resources in legacy packages are still
saved down using the version 4 format.