Add quick method support for lazy inst var access pattern
This provides a primitive response for methods of the following pattern:
```
instVar
    ^instVar ifNil: [...arbitrary code...]
```
This is common for lazy instance variable accessors, and will skip full
method activation if the instance variable is non-nil. It works by
providing a primitive that pushes the instance variable on the stack
if non-nil, but otherwise failing so that the Smalltalk code is run. The
code in the block does not necessarily need to initialize the variable.