Work in progress on correctly RenameClassRefactoring for namespace renames
When a namespace is renamed, it affects not just the renamed namespace,
but also all the classes in the namespace, and any references to those
classes.

This is a start on rewriting the references, although not working correctly
because the classes in the namespace still have the wrong names when the
outbound reference rewrite takes place. The cached fullnames of the
RBClasses need to be discarded and recomputed. These have also been used
to build a flat dictionary of all classes in the refactoring model. This
is a remnant of original design where there was a global namespace so a
single dictionary made sense. Removing it immediately was too big a change,
but it is probably time to dispense with it as it is redundant (the
RBNamespaceClass instances can be traversed instead to look up model
classes).