Maintain, not discard, Package loose method cache on repackage class
Rebuilding the loose method cache is fairly fast, but is still fairly
costly if performed repeatedly such as during the Package Browser tests.
To maintain it when moving a class to a new package, we just need to
remove any entries in the cache for methods of the class targetting
the new package, since such methods are no longer loose.

A number of other small optimisations are included that provide
incremental speed ups to PackageBrowserShellTests>>testRepackageClasses