Fix RemoveClassVariableRefactoring as not finding all references
This was broken in an earlier version of Dolphin when it became possible
to access class variables using a dot-qualified name. It is still not
correct in that we actually need to search the entire class hierarchy
for references.

Also improve the prompt that appears when there are references by
introducing a refactoring option for this similar to the other refactorings
that are now using this to display a richer message.