Replace most File class uses of CRT _splitpath
File class has a number of utilities for manipulating file pathnames.
These used _splitpath, but this isn't really intended for forking with
long pathnames.

Includes fix for bug in File class>>fullPathOf: when passed path
starting with .\. This is due to a bug in the underlying GetFullPathname
API, when fullPathOf: is invoked with a relative path that starts with
.\, the returned full path includes an extra pair of null terminators.