Fix renaming a class to a name shadowed by a temp in a referencing method
Renaming a class to the same name as a temp of instance variable in a
referencing method used to be treated as an error. We can now just add
extra qualification to the name (inst vars and temps cannot be qualified,
so can never clash with qualified names), but this was not being done.