Renaming class variables should also update any external references
For some time in Dolphin it has been possible to reference class variables
of classes outside the immediate hierarchy by using qualified names. The
Rename Class Variable refactoring has not updated these references as it
was not enhanced when the compiler change was made.

We can reuse much of the implementation of rewriting stale references to
static variables by pushing up the relevant functionality from the recently
updated Rename Class Refactoring.